Consolidate 3 Vercel functions into 1 to reduce Fluid compute costs
Merge api/mcp.ts, api/rest.ts into api/index.ts as a single entry point
with URL-based routing. Cache MCP server globally instead of recreating
per request. Centralize DB migration. Remove unused cron job.
